+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.1.0-alpha.1
+
+* Feature: Auto-install sentry-android SDK and integration dependencies (fragment, timber, okhttp) (#282)
+* Feature: `OkHttp` auto-instrumentation (#288)
+
 ## 3.0.0
 
 * Bump: AGP to 7.1.2 (#287)